and a few more
git-svn-id: http://svn.miranda-ng.org/main/trunk@11174 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/IcqOscarJ/src/icq_filerequests.cpp
@@ -124,7 +124,7 @@ void CIcqProto::handleFileRequest(PBYTE buf, size_t wLen, DWORD dwUin, DWORD dwC
 	wLen -= 4;
 
 	// Filename
-	WORD wFilenameLength;
+	size_t wFilenameLength;
 	unpackLEWord(&buf, &wFilenameLength);
 	if (!wFilenameLength) {
 		NetLog_Direct("Ignoring malformed file send request");
